+++ cmd2/utils.py
@@ -28,9 +28,10 @@ from typing import (
 from . import constants
 from . import string_utils as su
 from .types import (
-    ChoicesProviderUnbound,
     CmdOrSet,
-    CompleterUnbound,
+    CmdOrSetT,
+    UnboundChoicesProvider,
+    UnboundCompleter,
 )
 
 if TYPE_CHECKING:  # pragma: no cover
@@ -39,7 +40,7 @@ if TYPE_CHECKING:  # pragma: no cover
 else:
     PopenTextIO = subprocess.Popen
 
-_T = TypeVar('_T')
+T = TypeVar('T')
 
 
 def to_bool(val: Any) -> bool:
@@ -76,8 +77,8 @@ class Settable:
         settable_attrib_name: str | None = None,
         onchange_cb: Callable[[str, Any, Any], Any] | None = None,
         choices: Iterable[Any] | None = None,
-        choices_provider: ChoicesProviderUnbound[CmdOrSet] | None = None,
-        completer: CompleterUnbound[CmdOrSet] | None = None,
+        choices_provider: UnboundChoicesProvider[CmdOrSetT] | None = None,
+        completer: UnboundCompleter[CmdOrSetT] | None = None,
     ) -> None:
         """Settable Initializer.
 
@@ -185,7 +186,7 @@ def is_text_file(file_path: str) -> bool:
     return valid_text_file
 
 
-def remove_duplicates(items: Iterable[_T]) -> list[_T]:
+def remove_duplicates(items: Iterable[T]) -> list[T]:
     """Remove duplicates from an iterable while preserving order of the items.
 
     :param items: the items being pruned of duplicates